refactor: expriredTime 형식 변경 (Integer -> Long)

+++ src/main/java/gift/auth/service/dto/KakaoTokenResponse.java
@@ -12,13 +12,13 @@ public record KakaoTokenResponse(
         String tokenType,
 
         @JsonProperty("expires_in")
-        Integer expiresIn,
+        Long expiresIn,
 
         @JsonProperty("refresh_token")
         String refreshToken,
 
         @JsonProperty("refresh_token_expires_in")
-        Integer refreshTokenExpiresIn,
+        Long refreshTokenExpiresIn,
 
         @JsonProperty("scope")
         String scope
